// src/const/type/SlotTypeName.enum.ts
export enum SlotTypeName {
    chest = 'chest',
    chance = 'chance',
    go = 'go',
    visiting = 'visiting',
    banned = 'banned',
    tobanned = 'tobanned',
    free = 'free',
    lowtiertax = 'lowtiertax',
    lametax = 'lametax',
    utility = 'utility',
    station = 'station',
    brown = 'brown',
    lightblue = 'lightblue',
    pink = 'pink',
    orange = 'orange',
    red = 'red',
    yellow = 'yellow',
    green = 'green',
    blue = 'blue',
}